requirements.txt
flask==1.1.2
gunicorn==19.9.0
Pillow==8.2.0

